restructure cipher package
